network, driver: move the `DhcpConfig` unit tests next to its impl
