+++ pkg/network/driver/BUILD.bazel
@@ -33,11 +33,9 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//pkg/network/cache:go_default_library",
         "//vendor/github.com/coreos/go-iptables/iptables:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/github.com/onsi/ginkgo/extensions/table:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
-        "//vendor/github.com/vishvananda/netlink:go_default_library",
     ],
 )
